# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "منصة خبرات - Khibrat HR SaaS API",
        "description": "التوثيق الرسمي لكافة واجهات برمجة التطبيقات (APIs) لمنصة خبرات لإدارة الموارد البشرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سيرفر التطوير المحلي"
        }
    ],
    "paths": {
...
        "/api/companies/{company}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Retrieve company details with subscription info",
                "operationId": "0855706b525c08361e1128eb65ed7dc9",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company permanently",
                "operationId": "52cd05b2eb5827b6eebeafa0ee216d96",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted successfully"
                    }
                }
            }
        },
        "/api/companies/{company}/freeze": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Freeze a company's account",
                "operationId": "8f0db11fc07a533e570be1e7eaee9812",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company frozen successfully"
                    }
                }
            }
        },
        "/api/companies/{company}/activate": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Reactivate a company's account",
                "operationId": "ff05330a3f17f6bb53853e218ac594e2",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company reactivated successfully"
                    }
                }
            }
        },
        "/api/companies/stats": {
            "get": {
                "tags": [
                    "الشركات والاشتراكات (Companies & Subscriptions)"
                ],
                "summary": "جلب الإحصائيات الشاملة والمالية للوحة التحكم (Dashboard Stats)",
                "description": "يعيد هذا الـ API كافة الأرقام المالية، توزيع حالات الشركات للمخطط الدائري، البيانات الزمنية لآخر 6 أشهر للمخطط الخطي، وقائمة بآخر المنصات المسجلة لتغذية الـ Dashboard بالكامل.",
                "operationId": "71f7bc5257c0a017561ed9a80ad763cc",
                "responses": {
                    "200": {
                        "description": "تم جلب بيانات لوحة التحكم بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "properties": {
                                                        "total_revenue": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 2878
                                                        },
                                                        "monthly_revenue": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 49.99
                                                        },
                                                        "yearly_revenue": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 1200.5
                                                        },
                                                        "total_subscriptions": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "new_companies_this_month": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "auto_deletion_period_days": {
                                                            "type": "integer",
                                                            "example": 30
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "status_distribution": {
                                                    "properties": {
                                                        "active": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "frozen": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "at_risk": {
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "monthly_subscription_analytics": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "month": {
                                                                "type": "string",
                                                                "example": "Jun 2026"
                                                            },
                                                            "count": {
                                                                "type": "integer",
                                                                "example": 5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "latest_registered_platforms": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid",
                                                                "example": "74523827-0f69-45ce-be73-afd7be29f6e4"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "شركة النور التقنية"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2026-07-03"
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "active"
                                                            },
                                                            "package": {
                                                                "type": "string",
                                                                "example": "paid"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطأ داخلي في السيرفر أثناء احتساب الإحصائيات"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "أدخلي الـ Token الذي حصلتِ عليه من الـ Login هنا لتفعيل الصلاحيات",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}